// apps/web/src/shared/api/generated/model/loanHistoryControllerListResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * book-nest API
 * REST API for the book-nest project
 * OpenAPI spec version: 1.0
 */

export type LoanHistoryControllerListResult =
  (typeof LoanHistoryControllerListResult)[keyof typeof LoanHistoryControllerListResult];

export const LoanHistoryControllerListResult = {
  all: "all",
  on_time: "on_time",
  late: "late",
  no_due_date: "no_due_date",
} as const;