// apps/web/src/shared/api/generated/model/loanHistoryControllerListSort.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * book-nest API
 * REST API for the book-nest project
 * OpenAPI spec version: 1.0
 */

export type LoanHistoryControllerListSort =
  (typeof LoanHistoryControllerListSort)[keyof typeof LoanHistoryControllerListSort];

export const LoanHistoryControllerListSort = {
  returned_desc: "returned_desc",
  returned_asc: "returned_asc",
  loan_date_desc: "loan_date_desc",
  duration_desc: "duration_desc",
  title_asc: "title_asc",
  person_asc: "person_asc",
} as const;